# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": " App",
        "version": "v1.0.0"
    },
    "paths": {
...
        "/api/shop/{shop_id}/goods": {
            "get": {
                "tags": [
                    "Goods \u5546\u54c1\u7c7bApi"
                ],
                "summary": "\u5546\u94fa\u5546\u54c1\u5217\u8868",
                "description": "\u5546\u94fa\u5546\u54c1\u5217\u8868",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "shop_id",
                        "in": "path",
                        "description": "\u5546\u5e97id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            },
            "put": {
                "tags": [
                    "Goods \u5546\u54c1\u7c7bApi"
                ],
                "summary": "\u66f4\u65b0\u5546\u54c1",
                "description": "\u66f4\u65b0\u5546\u54c1",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u5546\u5e97id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "\u5546\u54c1\u540d\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "main_pic",
                        "in": "formData",
                        "description": "\u5546\u54c1\u4e3b\u56fe",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "media",
                        "in": "formData",
                        "description": "\u5546\u54c1\u5927\u56fe",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "\u5546\u54c1\u63cf\u8ff0",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "\u5546\u54c1\u8be6\u60c5",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "\u5546\u54c1\u72b6\u6001",
                        "required": true,
                        "type": "string",
                        "default": "\u6b63\u5e38"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "\u5546\u54c1\u7c7b\u578b",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "tag",
                        "in": "formData",
                        "description": "\u5546\u54c1\u6807\u7b7e",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "\u5546\u54c1\u4ef7\u683c",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "stock",
                        "in": "formData",
                        "description": "\u5546\u54c1\u5e93\u5b58",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "discount_price",
                        "in": "formData",
                        "description": "\u5546\u54c1\u6298\u6263\u4ef7\u683c",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            },
            "post": {
                "tags": [
                    "Goods \u5546\u54c1\u7c7bApi"
                ],
                "summary": "\u6dfb\u52a0\u5546\u54c1",
                "description": "\u6dfb\u52a0\u5546\u54c1",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u5546\u5e97id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "\u5546\u54c1\u540d\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "main_pic",
                        "in": "formData",
                        "description": "\u5546\u54c1\u4e3b\u56fe",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "big_pic",
                        "in": "formData",
                        "description": "\u5546\u54c1\u5927\u56fe",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "\u5546\u54c1\u63cf\u8ff0",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "\u5546\u54c1\u8be6\u60c5",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "\u5546\u54c1\u72b6\u6001",
                        "required": true,
                        "type": "string",
                        "default": "\u6b63\u5e38"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "\u5546\u54c1\u7c7b\u578b",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "tag",
                        "in": "formData",
                        "description": "\u5546\u54c1\u6807\u7b7e",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "\u5546\u54c1\u4ef7\u683c",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "stock",
                        "in": "formData",
                        "description": "\u5546\u54c1\u5e93\u5b58",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "discount_price",
                        "in": "formData",
                        "description": "\u5546\u54c1\u6298\u6263\u4ef7\u683c",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        },
        "/api/shop/{shop_id}/goods/{goods_id}": {
            "get": {
                "tags": [
                    "Goods \u5546\u54c1\u7c7bApi"
                ],
                "summary": "\u5546\u54c1\u8be6\u60c5",
                "description": "\u5546\u54c1\u8be6\u60c5",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "shop_id",
                        "in": "path",
                        "description": "\u5546\u5e97id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "goods_id",
                        "in": "path",
                        "description": "\u5546\u54c1id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Goods \u5546\u54c1\u7c7bApi"
                ],
                "summary": "\u5220\u9664\u5546\u54c1",
                "description": "\u5220\u9664\u5546\u54c1",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content_type",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "shop_id",
                        "in": "path",
                        "description": "\u5546\u5e97id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "goods_id",
                        "in": "path",
                        "description": "\u5546\u5e97id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        },
...
    },
...
}